add StopFlyingHandler
